- configured Tests (for VehicleRestController) - all dates are now LocalDateTime - added EventHistoryLog and corresponding views - added ApplicationSettings and corresponding view - fixed Patch requests - renamed EventController to VehicleEventController

+++ src/main/java/com/example/demo/api/rest_controllers/OrganisationRestController.java
@@ -1,11 +1,15 @@
 package com.example.demo.api.rest_controllers;
 
+import com.example.demo.database.models.EventHistoryLog;
 import com.example.demo.database.models.Organisation;
 import com.example.demo.database.models.utils.Mapping;
 import com.example.demo.database.models.utils.RestResponse;
 import com.example.demo.database.models.utils.ValidationResponse;
+import com.example.demo.database.services.EventHistoryLogService;
 import com.example.demo.database.services.OrganisationService;
-import com.example.demo.utils.StringUtils;
+import com.example.demo.utils.Constants;
+import com.example.demo.utils.DateUtils;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -16,24 +20,33 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.lang.reflect.Field;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(StringUtils.JSON_API + "/organisations")
+@RequestMapping(Constants.JSON_API + "/organisations")
 public class OrganisationRestController {
 
 	private final String ENTITY = "organisation";
 
+	@Autowired
+	private final EventHistoryLogService eventHistoryLogService;
 
 	@Autowired
 	private final OrganisationService organisationService;
 
 
+
 	@PostMapping(value = {"/batch"}, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<List<RestResponse<Organisation>>> postList(@RequestBody List<Organisation> organisations) {
+
+		if (organisations == null || organisations.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "NULL or empty array was provided");
+		}
 
 		boolean errorOccurred = false;
 
@@ -49,8 +62,6 @@ public class OrganisationRestController {
 				restResponse.setHttp_status(HttpStatus.BAD_REQUEST);
 				restResponse.setMessage(response.getMessage());
 
-				responseList.add(restResponse);
-
 				errorOccurred = true;
 			} else {
 				Organisation organisationFromDatabase = organisationService.save(organisation);
@@ -64,10 +75,12 @@ public class OrganisationRestController {
 					restResponse.setBody(organisationFromDatabase);
 					restResponse.setHttp_status(HttpStatus.OK);
 					restResponse.setMessage(ENTITY + " saved successfully");
-				}
 
-				responseList.add(restResponse);
+					addLog("create " + ENTITY, ENTITY + " created:\n" + organisationFromDatabase);
+				}
 			}
+
+			responseList.add(restResponse);
 		}
 
 		if (errorOccurred) {
@@ -76,7 +89,6 @@ public class OrganisationRestController {
 			return ResponseEntity.status(HttpStatus.OK).body(responseList);
 		}
 	}
-
 
 	@PostMapping(value = {"", "/"}, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<RestResponse<Organisation>> post(@RequestBody Organisation organisation) {
@@ -104,6 +116,8 @@ public class OrganisationRestController {
 			restResponse.setBody(organisationFromDatabase);
 			restResponse.setHttp_status(HttpStatus.OK);
 			restResponse.setMessage(ENTITY + " saved successfully");
+
+			addLog("create " + ENTITY, ENTITY + " created:\n" + organisationFromDatabase);
 
 			return ResponseEntity.status(HttpStatus.OK).body(restResponse);
 		}
@@ -137,6 +151,10 @@ public class OrganisationRestController {
 	@PutMapping(value = {"", "/"}, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<List<RestResponse<Organisation>>> putList(@RequestBody List<Organisation> organisations) {
 
+		if (organisations == null || organisations.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "NULL or empty array was provided");
+		}
+
 		boolean errorOccurred = false;
 
 		List<RestResponse<Organisation>> responseList = new ArrayList<>();
@@ -151,10 +169,9 @@ public class OrganisationRestController {
 				restResponse.setHttp_status(HttpStatus.BAD_REQUEST);
 				restResponse.setMessage(response.getMessage());
 
-				responseList.add(restResponse);
-
 				errorOccurred = true;
 			} else {
+				String oldOrganisationFromDatabase = organisationService.getById(organisation.getId()).toString();
 				Organisation organisationFromDatabase = organisationService.save(organisation);
 
 				if (organisationFromDatabase == null) {
@@ -166,10 +183,12 @@ public class OrganisationRestController {
 					restResponse.setBody(organisationFromDatabase);
 					restResponse.setHttp_status(HttpStatus.OK);
 					restResponse.setMessage(ENTITY + " saved successfully");
-				}
 
-				responseList.add(restResponse);
+					addLog("update (PUT) " + ENTITY, ENTITY + " updated from:\n" + oldOrganisationFromDatabase + "\nto:\n" + organisationFromDatabase);
+				}
 			}
+
+			responseList.add(restResponse);
 		}
 
 		if (errorOccurred) {
@@ -193,6 +212,7 @@ public class OrganisationRestController {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(restResponse);
 		}
 
+		String oldOrganisationFromDatabase = organisationService.getById(organisation.getId()).toString();
 		Organisation organisationFromDatabase = organisationService.save(organisation);
 
 		if (organisationFromDatabase == null) {
@@ -205,6 +225,8 @@ public class OrganisationRestController {
 			restResponse.setHttp_status(HttpStatus.OK);
 			restResponse.setMessage(ENTITY + " saved successfully");
 
+			addLog("update (PUT) " + ENTITY, ENTITY + " updated from:\n" + oldOrganisationFromDatabase + "\nto:\n" + organisationFromDatabase);
+
 			return ResponseEntity.status(HttpStatus.OK).body(restResponse);
 		}
 	}
@@ -214,15 +236,25 @@ public class OrganisationRestController {
 	@PatchMapping(value = {"", "/"}, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<List<RestResponse<?>>> patchList(@RequestBody List<Map<String, Object>> changesList) {
 
+		if (changesList == null || changesList.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "NULL or empty array was provided");
+		}
+
 		List<RestResponse<?>> responseList = new ArrayList<>();
 		boolean errorOccurred = false;
 
 		for (Map<String, Object> changes : changesList) {
 
-			changes.remove("password");
-
 			RestResponse<Map<String, Object>> mapResponse = new RestResponse<>();
 			mapResponse.setBody(changes);
+
+			if (changes == null) {
+				mapResponse.setHttp_status(HttpStatus.BAD_REQUEST);
+				mapResponse.setMessage("NULL array element was provided");
+				responseList.add(mapResponse);
+				errorOccurred = true;
+				continue;
+			}
 
 			if (!changes.containsKey("id")) {
 				mapResponse.setHttp_status(HttpStatus.BAD_REQUEST);
@@ -242,19 +274,11 @@ public class OrganisationRestController {
 
 					errorOccurred = true;
 				} else {
-					Integer id = (Integer) idObj;
-
-					Organisation organisationFromDatabase = organisationService.getById(Long.valueOf(id));
-
 					changes.remove("id");
+					long idLong = (long) ((Integer) idObj);
 
-					changes.forEach((key, value) -> {
-						Field field = ReflectionUtils.findField(Organisation.class, key);
-						if (field != null) {
-							field.setAccessible(true);
-							ReflectionUtils.setField(field, organisationFromDatabase, value);
-						}
-					});
+					String oldOrganisationFromDatabase = organisationService.getById(idLong).toString();
+					Organisation organisationFromDatabase = handlePatchChanges(idLong, changes);
 
 					RestResponse<Organisation> restResponse = new RestResponse<>();
 					restResponse.setBody(organisationFromDatabase);
@@ -267,8 +291,21 @@ public class OrganisationRestController {
 
 						errorOccurred = true;
 					} else {
-						restResponse.setHttp_status(HttpStatus.OK);
-						restResponse.setMessage(ENTITY + "patched successfully");
+
+						Organisation updatedOrganisationFromDatabase = organisationService.save(organisationFromDatabase);
+
+						if (updatedOrganisationFromDatabase == null) {
+							restResponse.setHttp_status(HttpStatus.INTERNAL_SERVER_ERROR);
+							restResponse.setMessage("failed to save " + ENTITY + " in database");
+
+							errorOccurred = true;
+						} else {
+							restResponse.setBody(updatedOrganisationFromDatabase);
+							restResponse.setHttp_status(HttpStatus.OK);
+							restResponse.setMessage(ENTITY + "patched successfully");
+
+							addLog("update (PATCH) " + ENTITY, ENTITY + " updated from:\n" + oldOrganisationFromDatabase + "\nto:\n" + updatedOrganisationFromDatabase);
+						}
 					}
 
 					responseList.add(restResponse);
@@ -286,29 +323,27 @@ public class OrganisationRestController {
 	@PatchMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<RestResponse<Organisation>> patchById(@RequestBody Map<String, Object> changes, @PathVariable Long id) {
 
+		if (changes == null || changes.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "NULL or empty array was provided");
+		}
+
 		Organisation organisationFromDatabase = organisationService.getById(id);
 
 		if (organisationFromDatabase == null) {
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, ENTITY + " with ID: '" + id + "' not found");
 		}
 
-		changes.remove("id");
-		changes.remove("password");
+		String oldOrganisation = organisationFromDatabase.toString();
 
-		changes.forEach((key, value) -> {
-			Field field = ReflectionUtils.findField(Organisation.class, key);
-			if (field != null) {
-				field.setAccessible(true);
-				ReflectionUtils.setField(field, organisationFromDatabase, value);
-			}
-		});
+		changes.remove("id");
+
+		organisationFromDatabase = handlePatchChanges(id, changes);
 
 		RestResponse<Organisation> restResponse = new RestResponse<>();
 		restResponse.setBody(organisationFromDatabase);
 		
 		ValidationResponse response = organisationService.validate(organisationFromDatabase, Mapping.PATCH);
 		
-
 		if (!response.isValid()) {
 			restResponse.setHttp_status(HttpStatus.BAD_REQUEST);
 			restResponse.setMessage(response.getMessage());
@@ -328,6 +363,8 @@ public class OrganisationRestController {
 			restResponse.setHttp_status(HttpStatus.OK);
 			restResponse.setMessage(ENTITY + " saved successfully");
 
+			addLog("update (PATCH) " + ENTITY, ENTITY + " updated from:\n" + oldOrganisation + "\nto:\n" + patchedOrganisation);
+
 			return ResponseEntity.status(HttpStatus.OK).body(restResponse);
 		}
 	}
@@ -336,6 +373,10 @@ public class OrganisationRestController {
 
 	@DeleteMapping(value = {"", "/"}, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<List<RestResponse<Organisation>>> deleteList(@RequestBody List<Organisation> organisations) {
+
+		if (organisations == null || organisations.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "NULL or empty array was provided");
+		}
 
 		boolean errorOccurred = false;
 
@@ -351,8 +392,6 @@ public class OrganisationRestController {
 				restResponse.setHttp_status(HttpStatus.BAD_REQUEST);
 				restResponse.setMessage(response.getMessage());
 
-				responseList.add(restResponse);
-
 				errorOccurred = true;
 			} else {
 				try {
@@ -360,15 +399,17 @@ public class OrganisationRestController {
 
 					restResponse.setHttp_status(HttpStatus.OK);
 					restResponse.setMessage(ENTITY + " deleted successfully");
+
+					addLog("delete " + ENTITY, ENTITY + " deleted:\n" + organisation);
 				} catch (Exception e) {
 					restResponse.setHttp_status(HttpStatus.INTERNAL_SERVER_ERROR);
 					restResponse.setMessage("failed to delete " + ENTITY + " from database \n" + e.getMessage());
 
 					errorOccurred = true;
 				}
-
-				responseList.add(restResponse);
 			}
+
+			responseList.add(restResponse);
 		}
 
 		if (errorOccurred) {
@@ -386,6 +427,12 @@ public class OrganisationRestController {
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, ENTITY + " with ID: '" + id + "' not found");
 		}
 
+		ValidationResponse response = organisationService.validate(organisationFromDatabase, Mapping.DELETE);
+
+		if (!response.isValid()) {
+			throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, response.getMessage());
+		}
+
 		try {
 			organisationService.delete(organisationFromDatabase);
 		} catch (Exception e) {
@@ -395,26 +442,47 @@ public class OrganisationRestController {
 		RestResponse<Organisation> restResponse = new RestResponse<>();
 		restResponse.setBody(organisationFromDatabase);
 		restResponse.setHttp_status(HttpStatus.OK);
-		restResponse.setMessage(ENTITY + "successfully deleted");
+		restResponse.setMessage(ENTITY + " deleted successfully");
+
+		addLog("delete " + ENTITY, ENTITY + " deleted:\n" + organisationFromDatabase);
 
 		return ResponseEntity.ok(restResponse);
 	}
 
 
+	private void addLog(String action, String description) {
+		if (eventHistoryLogService.isLoggingEnabledForOrganisations()) {
+			EventHistoryLog log = new EventHistoryLog();
+			log.setWho_did(eventHistoryLogService.getCurrentUser() == null ? "NULL" : eventHistoryLogService.getCurrentUser().toString());
+			log.setAction(action);
+			log.setDescription(description);
 
-	//TODO FOR TESTING -> REMOVE IN PRODUCTION
-	@PostMapping("/populate_with_test_data")
-	public void populateWithTestData() {
-		for (int i = 1; i < 11; i++) {
-			Organisation organisation = new Organisation();
-			organisation.setName("Organisation_" + i);
-			organisationService.save(organisation);
+			eventHistoryLogService.save(log);
 		}
 	}
 
-	//TODO FOR TESTING -> REMOVE IN PRODUCTION
-	@DeleteMapping("/delete_all")
-	public void deleteAll() {
-		organisationService.deleteAll();
+	private Organisation handlePatchChanges(Long id, Map<String, Object> changes) {
+		Organisation entity = organisationService.getById(id);
+
+		if (entity != null) {
+			changes.forEach((key, value) -> {
+				Field field = ReflectionUtils.findField(entity.getClass(), key);
+
+				if (field != null) {
+					field.setAccessible(true);
+
+					if (field.getType().equals(String.class)) {
+						ReflectionUtils.setField(field, entity, value);
+					} else {
+						if (field.getType().equals(Date.class)) {
+							LocalDateTime localDateTime = LocalDateTime.parse((String) value, DateUtils.getFormat());
+							ReflectionUtils.setField(field, entity, localDateTime);
+						}
+					}
+				}
+			});
+		}
+
+		return entity;
 	}
 }
